+++ core/common/src/main/java/hera/api/model/ContractTxHash.java
@@ -10,6 +10,7 @@ import hera.util.Adaptor;
 
 @ApiAudience.Public
 @ApiStability.Unstable
+@Deprecated
 public class ContractTxHash extends TxHash implements Adaptor, Encodable {
 
   /**