+++ core/transport/src/main/java/hera/client/ContractMethods.java
@@ -24,7 +24,6 @@ import hera.api.model.ContractDefinition;
 import hera.api.model.ContractInterface;
 import hera.api.model.ContractInvocation;
 import hera.api.model.ContractResult;
-import hera.api.model.ContractTxHash;
 import hera.api.model.ContractTxReceipt;
 import hera.api.model.Event;
 import hera.api.model.EventFilter;
@@ -92,12 +91,12 @@ class ContractMethods extends AbstractMethods {
 
         @Override
         protected void validate(final List<Object> parameters) {
-          validateType(parameters, 0, ContractTxHash.class);
+          validateType(parameters, 0, TxHash.class);
         }
 
         @Override
         protected ContractTxReceipt runInternal(final List<Object> parameters) throws Exception {
-          final ContractTxHash txHash = (ContractTxHash) parameters.get(0);
+          final TxHash txHash = (TxHash) parameters.get(0);
           logger.debug("Get receipt with txHash: {}", txHash);
 
           final Rpc.SingleBytes rpcDeployTxHash = Rpc.SingleBytes.newBuilder()
@@ -111,7 +110,7 @@ class ContractMethods extends AbstractMethods {
       };
 
   @Getter
-  protected final RequestMethod<ContractTxHash> deployTx = new RequestMethod<ContractTxHash>() {
+  protected final RequestMethod<TxHash> deployTx = new RequestMethod<TxHash>() {
 
     @Getter
     protected final String name = CONTRACT_DEPLOYTX;
@@ -125,7 +124,7 @@ class ContractMethods extends AbstractMethods {
     }
 
     @Override
-    protected ContractTxHash runInternal(final List<Object> parameters) throws Exception {
+    protected TxHash runInternal(final List<Object> parameters) throws Exception {
       final Signer signer = (Signer) parameters.get(0);
       final ContractDefinition contractDefinition = (ContractDefinition) parameters.get(1);
       final long nonce = (long) parameters.get(2);
@@ -141,14 +140,13 @@ class ContractMethods extends AbstractMethods {
           .fee(fee)
           .build();
       final Transaction signed = signer.sign(rawTransaction);
-      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed))
-          .adapt(ContractTxHash.class);
+      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed));
     }
 
   };
 
   @Getter
-  protected final RequestMethod<ContractTxHash> redeployTx = new RequestMethod<ContractTxHash>() {
+  protected final RequestMethod<TxHash> redeployTx = new RequestMethod<TxHash>() {
 
     @Getter
     protected final String name = CONTRACT_REDEPLOYTX;
@@ -163,7 +161,7 @@ class ContractMethods extends AbstractMethods {
     }
 
     @Override
-    protected ContractTxHash runInternal(final List<Object> parameters) throws Exception {
+    protected TxHash runInternal(final List<Object> parameters) throws Exception {
       final Signer signer = (Signer) parameters.get(0);
       final ContractAddress existingContract = (ContractAddress) parameters.get(1);
       final ContractDefinition contractDefinition = (ContractDefinition) parameters.get(2);
@@ -182,8 +180,7 @@ class ContractMethods extends AbstractMethods {
           .fee(fee)
           .build();
       final Transaction signed = signer.sign(rawTransaction);
-      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed))
-          .adapt(ContractTxHash.class);
+      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed));
     }
   };
 
@@ -224,7 +221,7 @@ class ContractMethods extends AbstractMethods {
       };
 
   @Getter
-  protected final RequestMethod<ContractTxHash> executeTx = new RequestMethod<ContractTxHash>() {
+  protected final RequestMethod<TxHash> executeTx = new RequestMethod<TxHash>() {
 
     @Getter
     protected final String name = CONTRACT_EXECUTETX;
@@ -238,7 +235,7 @@ class ContractMethods extends AbstractMethods {
     }
 
     @Override
-    protected ContractTxHash runInternal(final List<Object> parameters) throws Exception {
+    protected TxHash runInternal(final List<Object> parameters) throws Exception {
       final Signer signer = (Signer) parameters.get(0);
       final ContractInvocation contractInvocation = (ContractInvocation) parameters.get(1);
       final long nonce = (long) parameters.get(2);
@@ -254,8 +251,7 @@ class ContractMethods extends AbstractMethods {
           .fee(fee)
           .build();
       final Transaction signed = signer.sign(rawTransaction);
-      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed))
-          .adapt(ContractTxHash.class);
+      return transactionMethods.getCommit().invoke(Arrays.<Object>asList(signed));
     }
   };