+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Changes
   - Replace subscribeNewBlockXXX with subscribeBlockXXX
   - Clarify WalletApi, ContractApi
+  - Deprecate ContractTxHash
 
 ## v1.4.0 (Mar 10, 2020)